requirements-production.txt
aiofiles==23.2.1
aiohttp==3.12.6
anyio==4.9.0
crewai==0.121.1
fastapi==0.115.9
GitPython==3.1.44
httpx==0.28.1
mcp==1.9.3
openai==1.75.0
pydantic==2.11.5
python-dotenv==1.0.1
python-jose==3.5.0
python-multipart==0.0.20
redis==5.0.1
requests==2.32.3
sse-starlette==2.3.6
starlette==0.45.3
supabase==2.15.2
uvicorn==0.29.0
websockets==12.0
